Add LoRA training option
This adds a new `--lora_rank` arg, which if set to greater than 0 will
enable LoRA for our training instead of doing a full fine-tuning. It
defaults to 0, which preserves the existing default behavior.

This is useful for cases where there aren't enough GPUs available for
the full fine-tuning, but this deepspeed training script is still
preferred for something closer to the real training experience than
what `ilab train` gives.

Signed-off-by: Ben Browning <bbrownin@redhat.com>